% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_contents.R
\name{aggregate_gradients}
\alias{aggregate_gradients}
\title{Aggregate Gradient values}
\usage{
aggregate_gradients(all_gradient_files, iteration_number)
}
\arguments{
\item{all_gradient_files}{vector of character paths to files for
gradients to combine together on the computing site}

\item{iteration_number}{number of fitting iteration, used for tracking
and checking}
}
\value{
A list of estimated values, including the gradient,
sample size, iteration number, covariance matrix (\code{A_mat}),
number of samples with non-zero weights, the sum of the dispersion
values (for overdispersion estimates)
}
\description{
Aggregate Gradient values
}
