% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_available_models}
\alias{api_available_models}
\alias{api_get_current_beta}
\alias{api_model_specification}
\alias{api_submit_gradient}
\alias{api_model_converged}
\alias{api_setup_model}
\alias{api_clear_model}
\alias{api_estimate_model}
\title{API Functions and wrappers}
\usage{
api_available_models(url)

api_get_current_beta(url, model_name)

api_model_specification(url, model_name)

api_submit_gradient(
  url,
  model_name,
  data,
  site_name,
  shuffle_rows = TRUE,
  verbose = TRUE,
  dry_run = FALSE
)

api_model_converged(url, model_name)

api_setup_model(
  url,
  model_name,
  formula = "y ~ x1 + x2",
  family = "binomial",
  link = "logit",
  all_site_names
)

api_clear_model(url, model_name)

api_estimate_model(url, model_name, data, site_name, wait_time = 1, ...)
}
\arguments{
\item{url}{URL to the Plumber Server}

\item{model_name}{name of your model}

\item{data}{dataset to get gradient value from.  The code runs
\code{\link{gradient_value}} to calculate the gradient, no individual
data is submitted.}

\item{site_name}{name of the site, needs to be one of the
\code{all_site_names}}

\item{shuffle_rows}{should the rows of the dataset be permuted, so
as to decrease privacy concerns}

\item{verbose}{print out diagnostic messages}

\item{dry_run}{if \code{TRUE}, nothing with respect to the data
is submitted to the server, but returned to see what would be submitted.}

\item{formula}{model formula to fit, with tilde syntax}

\item{family}{generalized linear model family, see \code{\link{family}}}

\item{link}{link function to use with family}

\item{all_site_names}{all the site names to fit this model}

\item{wait_time}{Time, in seconds, to wait until to try to
get new estimate}

\item{...}{addition alarguments to send to
\code{\link{api_submit_gradient}}}
}
\description{
API Functions and wrappers
}
