% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{do_provision_glm_api}
\alias{do_provision_glm_api}
\alias{do_remove_glm_api}
\alias{do_deploy_glm_api}
\alias{do_deploy_glm_api_only}
\alias{do_list_plumber}
\title{Deploy GLM API on Digital Ocean (DO)}
\usage{
do_provision_glm_api(
  ...,
  application_name = "glm",
  port = 8000,
  example = FALSE,
  r_packages = NULL,
  github_r_packages = NULL
)

do_remove_glm_api(droplet, application_name = "glm")

do_deploy_glm_api(
  ...,
  application_name = "glm",
  port = 8000,
  swagger = TRUE,
  forward = TRUE,
  example = FALSE
)

do_deploy_glm_api_only(
  droplet,
  application_name = "glm",
  port = 8000,
  swagger = TRUE,
  forward = TRUE
)

do_list_plumber(droplet)
}
\arguments{
\item{...}{arguments to pass to \code{do_provision} from
\code{plumber} package}

\item{application_name}{Name of application, passed to \code{path}
argument of \code{do_deploy_api} function from \code{plumber} package}

\item{port}{port to deploy on Digital Ocean}

\item{example}{If TRUE, will deploy an example API
named hello to the server on port 8000.}

\item{r_packages}{Additional R packages to install, using
\code{install.packages}}

\item{github_r_packages}{Additional R packages to install from GitHub, using
\code{remotes::install_github}}

\item{droplet}{droplet to deploy on}

\item{swagger}{enable the Swagger interface,
passed to \code{do_deploy_api} function from
\code{plumber} package}

\item{forward}{setup requests targeting the root URL on the
server to point to this application,
passed to \code{do_deploy_api} function from
\code{plumber} package}
}
\value{
A droplet instance
}
\description{
Deploy GLM API on Digital Ocean (DO)
}
