% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{do_deploy_glm_api}
\alias{do_deploy_glm_api}
\title{Deploy GLM API on Digital Ocean (DO)}
\usage{
do_deploy_glm_api(
  ...,
  application_name = "glm",
  port = 8000,
  swagger = TRUE,
  forward = TRUE
)
}
\arguments{
\item{...}{arguments to pass to \code{do_provision} from
\code{plumber} package}

\item{application_name}{Name of application, passed to \code{path}
argument of \code{do_deploy_api} function from \code{plumber} package}

\item{port}{port to deploy on Digital Ocean}

\item{swagger}{enable the Swagger interface,
passed to \code{do_deploy_api} function from
\code{plumber} package}

\item{forward}{setup requests targeting the root URL on the
server to point to this application,
passed to \code{do_deploy_api} function from
\code{plumber} package}
}
\value{
A droplet instance
}
\description{
Deploy GLM API on Digital Ocean (DO)
}
