% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_contents.R
\name{estimate_new_beta}
\alias{estimate_new_beta}
\title{Estimate the updated beta value}
\usage{
estimate_new_beta(
  model_name,
  synced_folder,
  all_site_names = paste0("site", 1:3),
  tolerance = 1e-08
)
}
\arguments{
\item{model_name}{name of your model}

\item{synced_folder}{synced folder to do computation}

\item{all_site_names}{all the site names used to fit this model}

\item{tolerance}{tolerance for convergence}
}
\value{
A filename of the estimated values necessary for
the final estiamtes.
}
\description{
Estimate the updated beta value
}
