% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_contents.R
\name{estimate_site_gradient}
\alias{estimate_site_gradient}
\title{Estimate Site Gradient}
\usage{
estimate_site_gradient(
  model_name,
  synced_folder,
  site_name = "site1",
  data,
  all_site_names = paste0("site", 1:3),
  shuffle_rows = TRUE
)
}
\arguments{
\item{model_name}{name of your model}

\item{synced_folder}{synced folder to do computation}

\item{site_name}{name of the site, needs to be one of the
\code{all_site_names}}

\item{data}{dataset to get gradient value from}

\item{all_site_names}{all the site names used to fit this model}

\item{shuffle_rows}{should the rows of the dataset be permuted, so
as to decrease privacy concerns}
}
\value{
A character filename of the gradient file, with the
output from \code{\link{gradient_value}}
}
\description{
Estimate Site Gradient
}
