% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_contents.R
\name{gradient_value}
\alias{gradient_value}
\alias{use_glm_gradient_value}
\title{Estimate the update gradient value}
\usage{
gradient_value(
  beta = NULL,
  data,
  formula,
  family = binomial(),
  iteration_number = 0,
  shuffle_rows = TRUE
)

use_glm_gradient_value(
  beta = NULL,
  data,
  formula,
  family = binomial(),
  iteration_number = 0,
  shuffle_rows = TRUE
)
}
\arguments{
\item{beta}{current beta value, leave \code{NULL} to initialize}

\item{data}{dataset to get gradient value from}

\item{formula}{model formula to fit, with tilde syntax}

\item{family}{generalized linear model family, see \code{\link{family}}}

\item{iteration_number}{number of fitting iteration, used for tracking}

\item{shuffle_rows}{should the rows of the dataset be permuted, so
as to decrease privacy concerns}
}
\value{
A list of estimated values, including the gradient,
sample size, iteration number, covariance matrix (\code{A_mat}),
number of samples with non-zero weights, the sum of the dispersion
values (for overdispersion estimates), and a vector of values
for combining to create the population gradient (\code{u}), with length
of the number of beta values
}
\description{
Estimate the update gradient value
}
