% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_contents.R
\name{setup_model}
\alias{setup_model}
\title{Setup Model and Formula}
\usage{
setup_model(
  model_name,
  synced_folder,
  clear_model = TRUE,
  formula = y ~ x1 + x2,
  family = poisson()
)
}
\arguments{
\item{model_name}{name of your model}

\item{synced_folder}{synced folder to do computation}

\item{clear_model}{Should the model be cleared (all files deleted
model with same name) before creating new model}

\item{formula}{model formula to fit, with tilde syntax}

\item{family}{generalized linear model family, see \code{\link{family}}}
}
\value{
A character path to a formula/model file
}
\description{
Setup Model and Formula
}
